requirements.txt
facebook_sdk==3.1.0
markovify==0.8.0
requests==2.23.0
psaw==0.0.12
tweepy==3.8.0
random_username==1.0.2
beautifulsoup4==4.9.1
Pillow==7.2.0
